+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - #1155, #1602 AS3 remove trait button
 - #1260, #1438 AS1/2 direct editing on(xxx), onClipEvent(xxx) handlers
 - #1366, #1409, #1429, #1573, #1598 AS1/2/3 Add script/class (context menu on scripts folder)
+- Removing BUTTONCONDACTION, CLIPACTIONRECORD
 
 ### Fixed
 - #1298 AS1/2 properly decompiled setProperty/getProperty